feature outlet: using cleanup function from useEffet and OutletContext
