+++ src/router/index.tsx
@@ -1,42 +1,35 @@
 import { createBrowserRouter } from 'react-router-dom'
-import { useSelector } from 'react-redux'
-import Home from '@pages/Home'
 import ErrorPage from '@pages/ErrorPage'
 import BasketPage from '@pages/BasketPage'
 import EmptyBasketPage from '@pages/EmptyBasketPage'
-import Header from '@components/Header'
-import Wrapper from '@components/Wrapper'
 import LoginPage from '@pages/LoginPage'
+import Root from '@/pages/Root'
+import Home from '@/pages/Home'
 
 const router = createBrowserRouter([
   {
     // path: `/([ap].*)`,
-    path: '/*',
-    id: 'Homepage',
-    element: <Home />,
-    errorElement: <ErrorPage />
-  },
-  {
-    path: '/basket',
-    id: 'BasketPage',
-    element: (
-      <Wrapper>
-        <Header />
-        <BasketPage />
-      </Wrapper>
-    ),
-    errorElement: <ErrorPage />
-  },
-  {
-    path: '/emptyBasket',
-    id: 'EmptyBasketPage',
-    element: (
-      <Wrapper>
-        <Header />
-        <EmptyBasketPage />
-      </Wrapper>
-    ),
-    errorElement: <ErrorPage />
+    path: '/',
+    id: 'Root',
+    element: <Root />,
+    errorElement: <ErrorPage />,
+    children: [
+      {
+        path: '/',
+        id: 'Home',
+        element: <Home />
+      },
+      {
+        path: '/basket',
+        id: 'BasketPage',
+        element: <BasketPage />
+      },
+      {
+        path: '/emptyBasket',
+        id: 'EmptyBasketPage',
+        element: <EmptyBasketPage />
+      }
+    ]
   },
   {
     path: '/login',
